refactor(viewer): extract viewer_utils.jsx from viewer.jsx
Pure file split, no behavior change. Step 1 of mp-pxxc.
